refactor(bdd): deslop cli general-agent adapters
Dedupe the verdict-extraction block shared by both adapters into
toGeneralResult, share DEFAULT_TIMEOUT_MS and the temp-file path
construction in cli-agent.ts, inline the single-caller
findAuthFailureHint, drop the unused outputTail limit param, and
replace the sessionId `as string` cast with proper narrowing.
No behavior change.
